Extract mock objects into separate files
