+++ src/eventsMock.js
@@ -1,0 +1,28 @@
+const eventsMock = [
+  {
+    name: "Leeds Festival",
+    eventID: "312839",
+    location: "Leeds, United Kingdom",
+    startDate: "2018-08-25",
+    endDate: "2018-08-30",
+    maxAttendance: "150000"
+  },
+  {
+    name: "Reading Festival",
+    eventID: "312840",
+    location: "Reading, United Kingdom",
+    startDate: "2018-08-25",
+    endDate: "2018-08-30",
+    maxAttendance: "150000"
+  },
+  {
+    name: "Glastonbury",
+    eventID: "312841",
+    location: "Somerset, United Kingdom",
+    startDate: "2018-08-25",
+    endDate: "2018-08-30",
+    maxAttendance: "150000"
+  }
+];
+
+export default eventsMock;
